BUGFIX: Ensure workspace roles and metadata are pruned with the CR.
The `cr:prune` command is moved into Neos.Neos package where it has access to the workspaceService.
