+++ Neos.ContentRepositoryRegistry/Classes/Command/CrCommandController.php
@@ -13,6 +13,7 @@ use Neos\ContentRepositoryRegistry\Service\ProjectionReplayServiceFactory;
 use Neos\EventStore\Model\Event\SequenceNumber;
 use Neos\EventStore\Model\EventStore\StatusType;
 use Neos\Flow\Cli\CommandController;
+use Neos\Neos\Domain\Service\WorkspaceService;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Output\ConsoleOutput;
 use Symfony\Component\Console\Output\Output;
@@ -203,45 +204,5 @@ final class CrCommandController extends CommandController
             $progressBar->finish();
             $this->outputLine('<success>Done.</success>');
         }
-    }
-
-    /**
-     * This will completely prune the data of the specified content repository.
-     *
-     * @param string $contentRepository Name of the content repository where the data should be pruned from.
-     * @param bool $force Prune the cr without confirmation. This cannot be reverted!
-     * @return void
-     */
-    public function pruneCommand(string $contentRepository = 'default', bool $force = false): void
-    {
-        if (!$force && !$this->output->askConfirmation(sprintf('> This will prune your content repository "%s". Are you sure to proceed? (y/n) ', $contentRepository), false)) {
-            $this->outputLine('<comment>Abort.</comment>');
-            return;
-        }
-
-        $contentRepositoryId = ContentRepositoryId::fromString($contentRepository);
-
-        $contentStreamPruner = $this->contentRepositoryRegistry->buildService(
-            $contentRepositoryId,
-            new ContentStreamPrunerFactory()
-        );
-
-        $workspaceMaintenanceService = $this->contentRepositoryRegistry->buildService(
-            $contentRepositoryId,
-            new WorkspaceMaintenanceServiceFactory()
-        );
-
-        $projectionService = $this->contentRepositoryRegistry->buildService(
-            $contentRepositoryId,
-            $this->projectionServiceFactory
-        );
-
-        // reset the events table
-        $contentStreamPruner->pruneAll();
-        $workspaceMaintenanceService->pruneAll();
-        // reset the projections state
-        $projectionService->resetAllProjections();
-
-        $this->outputLine('<success>Done.</success>');
     }
 }
