+++ Neos.Neos/Classes/Command/CrCommandController.php
@@ -7,6 +7,8 @@ namespace Neos\Neos\Command;
 use League\Flysystem\Filesystem;
 use League\Flysystem\Local\LocalFilesystemAdapter;
 use Neos\ContentRepository\Core\Projection\CatchUpOptions;
+use Neos\ContentRepository\Core\Service\ContentStreamPrunerFactory;
+use Neos\ContentRepository\Core\Service\WorkspaceMaintenanceServiceFactory;
 use Neos\ContentRepository\Core\SharedModel\ContentRepository\ContentRepositoryId;
 use Neos\ContentRepository\Core\SharedModel\Workspace\ContentStreamId;
 use Neos\ContentRepository\Core\SharedModel\Workspace\WorkspaceName;
@@ -39,6 +41,7 @@ class CrCommandController extends CommandController
         private readonly ProjectionReplayServiceFactory $projectionReplayServiceFactory,
         private readonly AssetUsageService $assetUsageService,
         private readonly WorkspaceService $workspaceService,
+        private readonly ProjectionReplayServiceFactory $projectionServiceFactory,
     ) {
         parent::__construct();
     }
@@ -124,5 +127,50 @@ class CrCommandController extends CommandController
         $this->workspaceService->assignWorkspaceRole($contentRepositoryId, WorkspaceName::forLive(), WorkspaceRoleAssignment::createForGroup('Neos.Neos:LivePublisher', WorkspaceRole::COLLABORATOR));
 
         $this->outputLine('<success>Done</success>');
+    }
+
+    /**
+     * This will completely prune the data of the specified content repository.
+     *
+     * @param string $contentRepository Name of the content repository where the data should be pruned from.
+     * @param bool $force Prune the cr without confirmation. This cannot be reverted!
+     * @return void
+     */
+    public function pruneCommand(string $contentRepository = 'default', bool $force = false): void
+    {
+        if (!$force && !$this->output->askConfirmation(sprintf('> This will prune your content repository "%s". Are you sure to proceed? (y/n) ', $contentRepository), false)) {
+            $this->outputLine('<comment>Abort.</comment>');
+            return;
+        }
+
+        $contentRepositoryId = ContentRepositoryId::fromString($contentRepository);
+
+        $contentStreamPruner = $this->contentRepositoryRegistry->buildService(
+            $contentRepositoryId,
+            new ContentStreamPrunerFactory()
+        );
+
+        $workspaceMaintenanceService = $this->contentRepositoryRegistry->buildService(
+            $contentRepositoryId,
+            new WorkspaceMaintenanceServiceFactory()
+        );
+
+        $projectionService = $this->contentRepositoryRegistry->buildService(
+            $contentRepositoryId,
+            $this->projectionServiceFactory
+        );
+
+        // remove the workspace metadata and roles for this cr
+        $this->workspaceService->pruneRoleAsssignments($contentRepositoryId);
+        $this->workspaceService->pruneWorkspaceMetadata($contentRepositoryId);
+
+        // reset the events table
+        $contentStreamPruner->pruneAll();
+        $workspaceMaintenanceService->pruneAll();
+
+        // reset the projections state
+        $projectionService->resetAllProjections();
+
+        $this->outputLine('<success>Done.</success>');
     }
 }
